refactor: created SearchBar component for header
